core/requirements.txt
python-telegram-bot[rate-limiter]==20.1
openai>=0.27.0
tiktoken>=0.3.0
PyYAML==6.0
pymongo==4.3.3
python-dotenv==0.21.0
pydub==0.25.1
google-cloud==0.34.0
google-cloud-texttospeech==2.14.1
google-auth==2.5.0
fastapi==0.99.1
uvicorn[standard]==0.22.0
python-multipart==0.0.6
chromadb==0.4.8
pydevd-pycharm